+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Inefficient regular expression complexity of lowercase() and uppercase() rules
+
 ## [1.5.0] - 2022-06-27
 
 ### Added